requirements.txt
pyinstaller
qrcode
pillow